requirements.txt
dfpyre
pillow
pygame
pyopengl
pywebview